Add Tailwind CSS configuration file
Added a Tailwind CSS configuration file named 'tailwind.config.js' to the project root. This file specifies the content to be processed by Tailwind CSS, including HTML, JavaScript, TypeScript, and JSX files in the 'src' directory, as well as files from the 'flowbite-react' package. Additionally, the configuration extends the default theme and includes plugins for 'flowbite' and 'tailwind-scrollbar'.

Fixes: 3 (if applicable)