Add ChatSchema model
This commit adds the ChatSchema model to define the structure of chat documents in the database. The schema includes a 'members' field, which is an array containing the IDs of the members participating in the chat. The schema also includes timestamps to track when the chat was created and last updated. This model will be used to create, update, and retrieve chat documents in the database.